Partly done with building mechanic
I added some code to the HandleInteract() method so that you can now build when interacting with a "Build Area". All you have to do is mark the "Build Area" and the objects you're building with with either the "BuildRaft" or "BuildCar" tag and it should work. I just need to get the model of the raft and car and I can figure out the precise position of each piece.