Zombie AI so far
So far the zombie bean can follow the player if the scene is baked with the objects the zombie is supposed to avoid being in the Scenery tab. It follows the player and if it hits the player 10 times then the player "dies", it only puts a text on screen.